Acl redesign
Acl checks access for a role on a resource with a given privilege.

As role, the User is taken
As resource, 'any' other Entity is taken
As privilege, the applied Method is taken

The applied Method currently is SERVICE_NAME::METHOD_NAME.

But it could be any thing… also ENTITY_NAME::METHOD_NAME
or any thing else.