Implement loaded promise on objects retrieved from the API.
For objects in loading state, it is necessary to store the loaded promise in the Vuex store.
However, we do that without telling Vuex about it, so it doesn't mind and doesn't make the promise reactive.
An implication of this is that in the specs, we need to revert to expect(...).toMatchObject() instead of .toEqual(), in order to ignore the additional loaded properties in the store.