Don't attempt to reload entities that are already marked as deleting
Before this change, when deleting an entity which has a circular
reference to another entity (e.g. a camp and its owner, or a period and a
day in it), we would reload both of them indefinitely, always getting
status 404. We can avoid this by excluding entities from the reloading
if they are already marked as deleting.