backend/campcollaboration: delete entry if it was only invited or requested
Currently you have to delete an invitation if you mistyped the email.
Now the campcollaboration is really deleted on delete request.
Should be always like that, because the frontend should not try to delete an entry with the status established.
The backend should throw an error if the frontend tries to delete such an entry, not quietly update it.