Step 3: Update hal-json-vuex to 1.2.1
This version includes protection from redundant reload requests.

The activities needed on the activity detail view were eager-loaded in
step 2, but since they are also in a collection, a reload request for the
whole camp was sent out for each of them. The newest version of
hal-json-vuex avoids duplicate concurrent reload requests.