ECheckbox.spec.js: switch to explicit rendering
I think this improves the documentation because the component is
used as it would be in a view or another component.

Plus we don't have to use these input event hooks but
can verify the viewmodel of the outer component.

And it seems to change the rendering anyway, and i want
to test the rendering as it would be in a component/view.