Report coverage using GitHub Actions
Also, run backend tests with plain shell commands, because in the
Docker-based GitHub Action for phpunit, the project path is reported to
be under /github/workspace, while later for php-coveralls it is
/home/runner/work/ecamp3/ecamp3. This caused the file paths in the
clover XML coverage report to not match the files that php-coveralls
sees. So if we run phpunit outside of Docker, we keep using
/home/runner/work/ecamp3/ecamp3 everywhere, and it even runs a lot
faster.