Don't try to send a ci-passed event in forks
Forks don't have access to the main repo's secrets. When a branch on a
fork is pushed to, CI still runs, but the last step of sending out the
CI success event shouldn't happen there (unless the fork owner defines
his own REPO_ACCESS_TOKEN in the fork's secrets).

Unfortunately, GitHub Actions currently does not support checking the
presence of a secret in a conditional (see
https://github.com/actions/runner/issues/520). So for now, we just check
whether we are in the origin by checking the repo owner.