Prevent the weasy print worker from spamming Sentry
We didn't set a VHOST secret so far since we wanted to use the default
value of '/'. But as it turns out, because we still pass the nonexistent
secret to the deploy action, the action argument is not filled with the
default value. So we need to explicitly use the default values as
fallback.

As a consequence of the empty rabbitmq VHOST, the weasy print worker
went ahead and produced exceptions in a while: True loop. This busted
our sentry event limit overnight. So additionally, I set a timeout of 1
second after a failed connection attempt. In the puppeteer worker it was
already done this way.