Invitation: rewrite that the backend calculates all needed data
Also move the rejected screen to a separate route.
Before the InvitationController returned an embedded
camp, which the user had no access to.
(Because he was not in the camp)
When the user then tried to fetch the campCollaborations of the camp,
the title of the camp, available before,
was not available anymore.
This was an inconsistent behaviour, so i refactored it.

Issue: #631