Fix a cryptic Sentry issue in ApiWrapper
See https://sentry.io/organizations/ecamp/issues/2217764181/

ApiWrapper is able to write to normal fields as well as embedded
collections (e.g. in the case of an ApiSelect with multiple selection
activated). However, the check we used to find out whether the given
field is an embedded collection was flawed while the entity was still
loading from the API.

The ApiWrapper tests didn't catch this because so far, they have not
been completely honest in mocking API calls. This is fixed now, but is
yet another case where proper hal-json-vuex mocking capabilities would
help.