backend: switch to mysql:8.0.23 and limit the length of unique columns
Because we use MySQL for dev.ecamp3.ch.
In MySQL, you need to limit the indexed length of a
text or longtext column. But then the uniqueness is not
guaranteed on the whole length, so i just used the lengths
from before 0d98f35a31f345a7fed7a8e4facfd7f9889ab47e.

Issue: #1195