Fix namespacing issue in deployment (hopefully)
Locally it worked before, but when deploying, the classes in
eCampCore/data/dev could not read the classes in eCampCore/data/prod
(even though we declared them to be in the same namespace). This commit
tries to clean up the PSR-4 namespace structure somewhat, by making the
distinction between dev and prod data explicit in the namespace
structure.