CampCollaborationService: use random_bytes to generate inviteKey
Since 0b8204ca4d1663ce83bfbd506184b75427726444, if a dot (.) is in
the url and no slash is after that part of the path,
this part is interpreted as file an a 404 is returned by the vite dev server.
We did not test it in production.
We anyway wanted to use the same id generator for the inviteKey as for
the entities.

Issue: #1238