ci: use interpreter versions of docker-compose file
By using the development containers to execute the checks.
Also use php-cs-fixer version defined
in composer.json.

In a570e31f the php version in the docker-compose files was updated to 8,
but not the php version in the ci.
But the test wouldn't have passed with php v8.

In 3e925d7 the php-cs-fixer package was updated in
the composer.json, but the migration guide was not followed. The checks passed, because
the github action used the old version of
php-cs-fixer.