Refactor input filters to separate attribute from filtering logic
The reason for this is that the filtering logic might need some services
from the service container. The filter attribute is not instantiated
from the container, but rather via $attribute->newInstance() which is a
built-in PHP mechanism to get the instance of an attribute (including
its options). This mechanism of course does not support depdency
injection. The Symfony validators have the same structure of separated
annotation class and validator logic class, I presume for the same
reason.

Also similar to Symfony validators, we now have a filteredBy() method
that helps find the filter logic class responsible for a given filter
attribute. In order to instantiate this dynamically determined filter
logic class, we could use the serivce container, but that is discouraged
by Symfony. Rather, we use a service locator, which is the preferred
solution for this pattern: A service locator is a restricted container
which only has the services that are really needed from it. We configure
that service locator to include any services that match
App\InputFilter\*Filter.php, as well as any other services that are
tagged as 'app.input_filter'.

https://symfony.com/doc/current/service_container/service_subscribers_locators.html#defining-a-service-locator