Link related collections by overriding API platform serialization
By default, API platform will serialize linked collections by outputting
an array of links to the relation members. Instead, we would like it to
output a link to a collection, where these relation members can be
found.

With this modification to the HAL serializer, we get what we want iff a
relation satisfies the following conditions:
* The source entity defines the Doctrine relation with a targetEntity and
a mappedBy field
* The target entity has a SearchFilter for the reverse of the relation