Handle more edge cases in RelatedCollectionLinkNormalizer
For one, it is apparently possible for the decorated normalizer to
return a plain string IRI instead of a normalized object.

Additionally, for nested properties, the resource_class key in the
context is unset, so we use the actual class of the object for these
cases.

Thirdly, for now, we only support Doctrine associations. It is possible
to implement a custom PHP 8 Attribute that can be used to mark
non-associations with the necessary metadata, but for now, this is not
implemented.