Comment out unused code to prevent lowering the code coverage too much
The EagerLoadingExtension is only temporarily in our codebase, until
https://github.com/api-platform/core/pull/4377 is merged and released.

There are tests for this extension in API Platform core, so we shouldn't
need to write our own test suite for it. However, because of our limited
use cases so far, this class lowers the coverage too much. Therefore, I
commented out some code of which I am certain we won't need it in the
near future. Less untested code = higher coverage.