Automatically copy graphql operation security from rest
Security rules in REST and GraphQL are not correlated:
https://api-platform.com/docs/core/graphql/#security

Caution:
The operation security rules are not enough in GraphQL: We also need
to add security rules to all association ApiProperties in order to
prevent traversal attacks (see
https://github.com/api-platform/docs/pull/1338/files). This will be
available starting in API Platform 2.7. An example for a traversal attack:

{
    camp(id: "/camps/1a2b3c4d") {
        campCollaborations {
            user {
                id
                name
                email
                campCollaborations {
                    camp {
                        id
                        name
                        ...
                    }
                }
            }
        }
    }
}
Even though the current user might have the rights to access the single
camp, he probably doesn't have the right to access the other user's
emails and definitely not their campCollaborations (this is assuming we
exposed the user's campCollaborations some time in the future).