Set the JS-accessible cookie on the parent domain
This is meant as a temporary solution so at least our own API client
services can access the information in the JWT token. We still plan on
implementing a logout endpoint that will unset the JWT cookies and a
login status flag on the API root, so that external API clients can
check whether or not they are logged in.