CampProgram: allow creating activities
With the changes made with api platform.
Also use period iri instead of id to filter activies in Picasso.vue.
Remove periodId from the payload and send period iri in DialogActivityCreate.vue.
Use category iri to reference the category in DialogActivityForm.vue.

Use cascade persist in the Activity Entity to persist created ScheduleEntries.
Checked with docker-compose exec php     bin/console doctrine:migrations:diff that no migration is necessary.
ScheduleEntry already has onDelete=cascade on the joincolumn.

Use $this->activity?->getCamp() instead of property access in ScheduleEntry
because when an activity is created, the camp is null (this is attached in the persister),
and Activity::getCamp returns the camp of the category.

Issue: #2029