api: add /profiles endpoint
The /users endpoint is now exposed to all users and
exposes all properties which may be seen by all authenticated users.
Currently, this is only the displayName.
The /profiles endpoint is exposed to related users,
with all properties a related user may see.
Patch is only allowed for the /profiles endpoint, where
the user can edit the properties exposed to related users.
displayName is currently a read only property.

Use the strict equality operator for 'object === user',
because an exception was thrown for too deep recursion
with the == operator.

Issue: #2148