api: ensure that the /invitations endpoint is shown in /index.jsonhal
For /invitations, we need to add a dummy get collection operation.
For /invitations{/id}, we also need to define a get operation.
We cannot use a dummy get operation,
because then the self links would look like this:
'self' =>
     array (
-      'href' => '/invitations/myInviteKey2/find',
+      'href' => '/invitations/myInviteKey2',
     ),
If parameters with different names had been defined
in the same path segment, it would not make sense
to expose all the possible names.
Thus, we use the existing functionality of the UriTemplateFactory
to add an {/id} template parameter if the get item operation is defined.

fixup! api: ensure that the /invitations endpoint is shown in /index.jsonhal