InputFilterDenormalizer: also filter 1:1 related entities in the body
That is needed that it's possible to create a profile when
creating a user.
Else the InputFilter on the profile properties are not applied.

Extract new functionality to separate method.
Rename previous function getInputFilterAttributes to createInputFilterAttributesFrom,
because it's a factory method to create the InputFilter instances.
Extract getFilterAttributes method (which now gets the filter attributes from the whole class).

Issue: #2148