CampCollaboration: add unique constraints on (user, camp) and (inviteEmail, camp)
That it's not possible to have the same user twice in a camp
(which leads to further problems), and to use the same
inviteEmail twice in the same camp.
Because we use postgres now, we can use a normal unique index, because postgres does not consider null values for
the unique constraints.

Because the it's not possible now to invite another user with a suer reference (because a user cannot see another user),
some tests needed to be changed, and one needed to be skipped.