Set appVersion before deploying
This is needed to check which deployments need to be updated. Since we
run our deployment job on a schedule instead of based on pull request
events (because then secrets wouldn't be available to PRs from forks),
we need to perform some extra steps to make sure we don't deploy
unnecessarily all the time.