ci: switch back to running tests on the github runner
Instead of in the docker container.
Leave the tests for the old backend as they are, because
there we cannot and don't want to update to php 8.
We can delete these ci jobs later.
Use renovate regex matchers to update php, node and composer
in the docker containers and in the github actions.