Camp: assert at least one period in camp
It looks a little hacky, the reason is:
For a delete request, the validations are normally not processed
The camp looses the period only after the delete operation, not before.
If we remove the period from the camp before running the validation, the #[Assert\Valid] is not processes, because the camp property on period is null.
But if we want to delete 1 Period from a camp, there has to be at least one other,
so i added this validation on Camp.

Issue: #2405