Get it to render another font than Roboto
This was a pain to acheive, and I found no really good solution.
The only way that worked for both development and production, and for
both preview and pdf rendering, was to install the fonts into the
container, which blocks us from using the browserless.io service, and at
the same time specifying the Google Fonts URLs in the CSS.

Hosting the fonts locally in our nuxt server wouldn't work because
pagedjs couldn't parse the relative font URLs in the CSS, but only while
rendering the pdf, not if previewing with pagedjs in the browser.