POC pdf rendering in web worker
Only works in Chromium-based browsers, not in Firefox. For Firefox
support, we need to add https://www.npmjs.com/package/vite-plugin-worker

Currently, i18n and hal-json-vuex don't work, because the worker thread
does not have access to the memory and objects of the main thread. So we
need to pass the translations and vuex store contents to the worker
thread and construct instances of vue-i18n and hal-json-vuex there.