Refactor to only import React when necessary
React and other rendering dependencies are now only imported when
absolutely necessary. If rendering the PDF in the main thread, this
means React is imported via a dynamic import. If rendering in a web
worker, React is imported statically in the worker script, but the
worker script itself is imported dynamically.

Also introduces a clearer directory structure. "documents" are like
"views" in HTML rendering, i.e. root components which may contain other
components from the "components" directory.