Add sentry to worker and fix firefox cloning issue
If an error happened in the web worker, this error would be serialized
as an error object into the result returned from the web worker. Firefox
did not like this, while the other browsers were silent about it. But
really, only plain JSON serializable data may be passed between worker
and main thread (although blobs seem to be okay).
So instead, we use Sentry to log any error details in the worker, and
pass only the message string back to the main thread, to be logged or
displayed to the user.