Fix 'Unknown font format' error in production
Vite would previously simply omit the font files when bundling for
production. When react-pdf went on to fetch the fonts, vite would then
serve the index.html file instead of the font files. This led to the
Unknown font format error.
The simple fix in this commit forces vite to bundle the font files
during npm run build.