MaterialList: make column name nullable again
And align the schema migrations with the entities.

In 24ebe7a (#2452), the nullable:false was removed in assumption that the default is true.
(That's how it is in hibernate).
The schema migration set the field to nullable, but the entity did not.
Also add the unique index on campCollaborationId needed for the OneToOne relationship.
(i don't know how that got missing).

Issue: #2591