ci: fix coverage reports
the coverage reports with node + jest und with php are different.

node + jest previously reported the path to the files relative to pwd.
With the new version (v28), it reports the path relative to package.json,
which makes sense.
But coveralls.io cannot handle this, thus we need to report the file paths
from the root of the git repository.
Include all js and vue files in the coverage report for print
(except node_modules and .nuxt).
Remove the use of <rootDir> in collectCoverageFrom in print/package.json.

phpunit still reports the path relative to pwd were the tests were executed,
but you cannot define different prefixes for the files for php and javascript.