Fix vanishing text on short schedule entries
This involved setting the time columns to a carved-in-stone constant
width. Otherwise, as soon as the time column has some text contents, the
layouting engine would reserve only 7.199pt of height for the title of
some short schedule entries (even regardless of page height) for some
reason. Since the font size is set to 8, even the first "line" of the
schedule entry title text would then be completely omitted from rendering.

We fix the time column width for eternity by setting the width manually
and wrapping the contents in a position: absolute View.
The fixed width is calculated using a heuristic, based on the presumably
longest time displayed in the column. This heuristic could be greatly
improved if we used fontkit to actually measure the width of the string
as it will be rendered, but then we'd have to re-implement some of the
cascading styles logic ourselves... So for the moment, this simple
heuristic will have to do.