Use more mocking instead of allowing to override final classes
If a package vendor specifies a class as final, they really do not
expect the class to ever be overridden. Also, adding this package
to require-dev allows us to do so in a dev environment, even
accidentally outside of tests, and that code will only fail when
running in production.
Instead of trying to enable mocking SearchFilter, we can just
instantiate it and mock a little more in another place.
The test testNormalizeDoesntReplaceWhenStrategyIsNotExact was flawed,
because the search strategy is always forced to be "exact" when
filtering by an association. For this reason, I removed that test,
because it does not make sense to test a primitive field filter in the
context of RelatedCollectionLinkNormalizer. This normalizer is only
concerned with associations, not with primitive fields.