Switch to lockfileVersion 3
Signed off by @BacLuc and @carlobeltrame
According to
https://docs.npmjs.com/cli/v8/configuring-npm/package-lock-json?v=true#lockfileversion
the only difference between version 2 and 3 is that v3 omits some
backward compatibility affordances, namely the materialized transitive
dependency tree. The versions of all transitive dependencies are still
recorded in the lockfile, but in v2 it was redundantly also stored in
tree form. This tree form is now not needed anymore, because we are not
going to switch back to lockfile version 1 anytime soon. So we can
safely upgrade to v3.

This will require all open PRs which touched the frontend or print
dependencies to be rebased or to merge devel with an `npm install` for
resolving the merge conflicts.