Ignore twemoji files in prettier on CI
Apparently, the version of git which is installed on GitHub Actions
runners by default is too old for the very specific features we need for
cloning only the specific twemoji image files and nothing else. Oh well.
Everywhere else, we are using a modern enough version of git inside
Docker, so this isn't a big problem. Therefore we can just ignore the
code style of these cloned files on CI.