Fix dayjs locale in react print on deployments
For some reason, only in deployments, the dayjs copy used in the worker
is separate from the copy used in the main thread. So when starting a
render job, we have to set the locale again. Setting this once during
worker job initialization will then apply to the whole worker job.