continuous-integration.yml: use COVERALLS_TOKEN to send coveralls reports
This is how it is documented in https://github.com/nickmerwin/node-coveralls#usage
I dont know why this worked before.

For php we use the github action, and there we only need this special token for sending local coverage info to coveralls:
https://github.com/marketplace/actions/coveralls-github-action
For the github action, the GITHUB_ACTION token suffices.