e2e: move downloads folder instead of clearing it
That we can see the pdf generated.

Remove setting the language before the test.
This breaks the tests for firefox when you use tasks. Then the tasks don't get executed.
If you start to use beforeEach in combination with it, the tests fail directly with the error:
"No handler specified for remote:debugger:protocol"
In the container we anyway have only english.