.helm: add feature to protect application with basic auth
We need to send the 'Authorization' header also from print because
it uses the external url which is protected.

Using the kubernetes secret type https://kubernetes.io/docs/concepts/configuration/secret/#basic-authentication-secret
did not work, it did not generate a secret with $username:md5($password), but only had the 2
entries username and password.

Issue: #3288