print: also log fetch errors which will then be sent to sentry
They are logged in development mode, but not in production mode by default.
Found the hooks by searching for "callHook("" in node_modules/@nuxt,
and this hook was called in production mode after the fetch errors happened
with the fetch errors in the property.
This also helped: https://github.com/nuxt/nuxt/issues/7742#issuecomment-743205834