Fix off-by-one error which resulted in wrong positioning of schedule entries at day start
Either of the two changes would fix the noticed problem on their own.

The >= vs > fix is an off-by one error, which caused the percentage
algorithm to calculate a remainder which was larger than necessary.

The division change fixes a mistake, due to which the calculated
remainder would ignore the local weight scaling.

Due to the first fix, the remainder is now zero in the edge case, and
due to the second fix, the remainder is now scaled correctly (but it
does not matter anymore in the problematic edge case because then the
remainder is zero).