api/migrations/schema: keep extension pgcrypto added
We need it to add it to invite ourselves to the interesting camps on staging.
Adding the extension without IF EXISTS in Version20230409164830.php if it already exists leads to an error.
The modification of Version20230409164830.php fixes the backups which did not already execute this migration.
The migration Version20230731081633.php fixes the backups which already executed Version20230409164830.php.

Issue: #3478