CategoryCreateProcessor: set timestampable properties manually
In production this does not work in this case now.
Because reproduction and fix may be hard, we use this quickfix here.
This allows us to release and deploy in the meantime.
Use a method name which does not match the naming conventions for
the setters and getters, else the fields are null when creating a ColumnLayout directly.

Issue: #3662