sentry: add sourcemaps to frontend and print
Add .cache/npm/sentry-cli to cache.
Add it with a .gitkeep file that the permissions are set correctly.
Move the .gitkeep instruction from e2e up to the top level .gitignore.
It may be that you need to delete your node_modules folders for frontend and print after this commit.