api: set report_fields_where_declared:true in doctrine.yaml for doctrine/orm 3
And don't overwrite the doctrine definition for the same field
again with the same values.
Now we define the Constraints on the getter,
because api-platform only allows to define constraints on a getter.
Move the example on ContentNode::data to the getter, because then this example
was used for all children for the data property instead of the example
on the getter.

Issue: #3740