api: add snapshot tests for the api responses
- test the root endpoint without escaping the values
- test GetCollection endpoints with escaping the values
- test GetItem endpoints with escaping the values
Here using a map to use a fixture per collection endpoint was necessary,
because taking the first item of the collection endpoint would result in a different
item used which may have another structure. (e.g. more children or more related entities)
- test that Patch and GetItem result in the same response
This was not possible for the /profile endpoint because there some properties are present in
_links, in PATCH, but not in _embedded, but they still lead to a compatible response.