ApiWrapper.vue: fix autosave
In commit 24dc4bf the debouncedSave method was moved from
computed to the methods, where it seems to belong.
But the lodash debounce function returns a function, which
delays the invocation of the func when called, but does not directly
schedule the delayed function execution.
The caller has to do that by calling the returned function.
This apparently was done by vue before commit 24dc4bf,
because it was a computed property.

closes #3839