frontend: fix mobile view of materialLists
Navigate to the list of materiallists in the NavBottombar.
Add new view for the list of materiallists
to navigate to the details of one materiallist.
Rename the views/admin/(,Admin)MaterialLists component to avoid name clash with
the new component.
Add quick way to differentiate the target routes in the materialListRoute helper
for the NavBottombar and NavTopbar.
May have been introduced in #3841.

Also add guard if camp is not loaded to materialListRoute.