chore(deps): update @vue/runtime-core to 3.3.7
And also perform lock file maintenance.
The update of @vue/runtime-core to 3.3.5 did not play well
without updating the other dependencies.
Resolving vue components did not work anymore with the warning:
resolveComponent can only be used in render() or setup()
And then the vue components were not resolved which lead
the custom renderer to be called with the vue components, which
it could not handle.
(Tag <SimpleComponent> cannot be used inside a pdf. Did you forget to import a Vue component?)