print: pin @vue dependencies, that they have the same version
There seem to be problems on version mismatches.
https://github.com/ecamp/ecamp3/pull/4046
https://github.com/ecamp/ecamp3/pull/3937

For print specifically: https://github.com/ecamp/ecamp3/pull/4056

The transitive dependencies seem to be updated in the lock file maintenance.
The direct dependency in a normal package update.
Even in minor versions this seems to be a problem ->
we pin them together, and they get updated together.