frontend: load all dayResponsibles in one request per period Dashboard.vue
By loading the days in one request per period which embed the dayResponsibles.
When loading the dayResponsibles directly, there was still
a request per day for the dayResponsibles.
This makes the loading of Dashboard.vue faster.
It also may reduce the e2e test failures of the firefox tests due to:
"Request aborted" errors after the test is finished.
(The test did not wait for the dayResponsibles to be loaded and then
aborted the load which lead to the requests being aborted.)