EndpointQueryCountTest: allow a range for the number of queries on content_node endpoints
With the ResponsiveLayout, we now have mostly ContentNodes which are nested by 2 levels:
ColumnLayout -> DefaultLayout -> "Content Nodes with content"
This may lead to n+1 db queries that we experienced in the EndpointQueryCountTest.
This change now allows to define a range for the query count.
This is good enough to keep the performance at the same level.