common: do not use dateHelperVCalendar.js anymore in picasso.js
In the picasso render logic in pdf and in print, we can use the utc dayjs objects directly, instead
of converting them to the timestamps which represent the same time in the local timezone.
This fixes the render bugs of the picasso in the pdf module if you render it in another timezone.
Changed functions:
- positionStyles: only used in pdf and print
- filterScheduleEntriesByDay: ony used in pdf and print

Also remove the other usages in pdf (DayColumn) and print (CalendarDay).