Remove unnecessary sorting
We don't need to sort because all schedule entries of the period are
fetched when printing is started, and the schedule entries of a period
are ordered correctly when fetching from the API. This might become
relevant again far in the future if we introduce more offline
functionality.

For now, the sorting was incorrect anyways, because it assumed that the
scheduleEntryNumber was unique per day, which it isn't (there can be
multiple numbering schemes leading to multiple of the same
scheduleEntryNumbers on the same day).