api: add environment performance_test
To test our application against more data.
This uses the feature of the AliceBundle that
the fixtures can be loaded depending on the environment.
(https://github.com/theofidry/AliceBundle/blob/ab4d7ffd04eef576df4af88630ebb49bba397e40/doc/advanced-usage.md#environment-specific-fixtures)

We need that because the nelmio/alice takes a while to generate lots of data,
thus we don't want to do this for every test.

We need a database with more data to test our application against
a postgres instance which needs to use its indices.
Else it will just do a full table scan every time and this is fine
with 5 entries in the table.
But if it then reaches to an inefficient query plan as soon as
we have a lot of data, we face issues in production.
This allows us to test that before it happens in production.

The amount of additional data in additional camps has to be balanced.
Now it takes 50s on a fairly good machine to load all these fixtures.
If we later also add more activities to camp1, we cannot load
a lot more additional camps.
With 1000 additional camps and 100 additional activities for camp1
it took 15 min to load the fixtures.