print: migrate to flat config
Using npx @eslint/migrate-config .eslintrc.json
Then add the additional packages needed
(@eslint/compat, globals, @eslint/eslintrc, @eslint/js)
Add paths from the .gitignore file to the ignored patterns.
Also lint .ts files with eslint (this seems not to be the default).
Remove the options from the call to eslint which are not
supported anymore in the package.json.
Ignore common, .nuxt and .output explicitly, because it did not work with
includeIgnoreFile.

Issue: #5282