replace uri-js with uri-js-replace
That we have less dependencies on punycode.
uri-js is not maintained anymore, thus we won't get an update.
uri-js-replace removed the dependency on punycode.

This starts to solve the deprecation warning:
"punycode is deprecated, use a userland alternative instead."

Issue: #5843

Can be reverted as soon as the dependencies don't depend on
uri-js anymore.