httpCache.cy.js: fix unstable 'invalidates /camp/{campId}/categories for new category'
In cypress the overall model for assertions is
auto waiting, and this works well enough by just
using sequential statements.
But for requests where auto retry or waiting is impossible or even
maybe even dangerous, you need to model the dependencies directly with "then".

expectCacheMiss is also a special assertion because it uses
a request (it's GET, a retry could be implemented), but we want
the cache invalidation to be synchronous, und thus want
to test the cache logic after the login request is completed.
(And not wait until the cache status switches to MISS for one request).

Issue: #5493